form-flow-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
playwright>=1.48.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
google-generativeai>=0.8.0
openai>=1.54.0
pydantic>=2.10.0
requests>=2.32.0
aiohttp>=3.10.0
tornado>=6.0.3,<7
branca
folium>=0.13,!=0.15.0
matplotlib>=3.3.0
psutil
py-cpuinfo
seaborn>=0.11.0
elevenlabs==1.6.1